# services/gateway/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Interactive documentation for Order \u0026 Customer services",
        "title": "Customer \u0026 Order gRPC services",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/customer/": {
            "get": {
                "description": "Returns all customers as json array",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customer"
                ],
                "summary": "Returns all customers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.Customer"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new customer and returns the customer's UUIDv4",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customer"
                ],
                "summary": "Creates a new customer",
                "parameters": [
                    {
                        "description": "Customer info",
                        "name": "customer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.CustomerRequestCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/customer/validate/{UUIDv4}": {
            "get": {
                "description": "Checks if the customer is in the database and returns boolean",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customer"
                ],
                "summary": "Checks if the customer is in the database",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Customer UUIDv4",
                        "name": "UUIDv4",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/customer/{UUIDv4}": {
            "get": {
                "description": "Returns customer data as json with given UUIDv4",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customer"
                ],
                "summary": "Returns customer data with given UUIDv4",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Customer UUIDv4",
                        "name": "UUIDv4",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.Customer"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates customer's info and returns boolean",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customer"
                ],
                "summary": "Updates customer's info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Customer UUIDv4",
                        "name": "UUIDv4",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Customer info",
                        "name": "customer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.CustomerRequestUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes the customer and returns boolean",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customer"
                ],
                "summary": "Deletes the customer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Customer UUIDv4",
                        "name": "UUIDv4",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "types.Address": {
            "type": "object",
            "properties": {
                "addressLine": {
                    "type": "string",
                    "example": "x street, y apartment, number:5/7"
                },
                "city": {
                    "type": "string",
                    "example": "New york"
                },
                "cityCode": {
                    "type": "integer",
                    "example": 45
                },
                "country": {
                    "type": "string",
                    "example": "United States"
                }
            }
        },
        "types.Customer": {
            "type": "object",
            "properties": {
                "address": {
                    "$ref": "#/definitions/types.Address"
                },
                "createdAt": {
                    "type": "string",
                    "example": "2022-05-23T09:45:01.675884703Z"
                },
                "customerUUID": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string",
                    "example": "2022-05-23T09:55:04.995285414Z"
                }
            }
        },
        "types.CustomerRequestCreate": {
            "type": "object",
            "properties": {
                "address": {
                    "$ref": "#/definitions/types.Address"
                },
                "email": {
                    "type": "string",
                    "example": "johndoe@example.com"
                },
                "name": {
                    "type": "string",
                    "example": "John Doe"
                }
            }
        },
        "types.CustomerRequestUpdate": {
            "type": "object",
            "properties": {
                "address": {
                    "$ref": "#/definitions/types.Address"
                },
                "email": {
                    "type": "string",
                    "example": "johndoe@example.com"
                },
                "name": {
                    "type": "string",
                    "example": "John Doe"
                }
            }
        },
...
    }
}